% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-plot.R
\name{convert_to_code}
\alias{convert_to_code}
\title{Reverse plot}
\usage{
convert_to_code(p, ggplot_data_name = "ggplot_data_name",
  layer_data_names = NULL)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{ggplot_data_name}{name of primary data source. An attempt is made to
find the actual name of the data variable by comparing the data stored
in the plot object against all variables in the object's \code{plot_env}.
If no matching data is found, then \code{ggplot_data_name} is used as
the name of the data variable.}

\item{layer_data_names}{name of layer data sources. If a layer simply inherits
the root data source, then no name is needed. Otherwise, an attempt is made to
find the actual name of the data variable by comparing the data stored
in the plot layer against all variables in the plot's \code{plot_env}.
If no matching data is found, then \code{layer_data_name} is used as
the name of the data variable.}
}
\value{
text to create the given plot
}
\description{
Reverse plot
}
