% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fargs}
\alias{fargs}
\title{Format named list into argument text for a function call}
\usage{
fargs(nlist)
}
\arguments{
\item{nlist}{named list}
}
\value{
single character string
}
\description{
Format named list into argument text for a function call
}
