% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_complete_theme}
\alias{reverse_complete_theme}
\title{Convert an complete \code{theme} into code to create it}
\usage{
reverse_complete_theme(theme)
}
\arguments{
\item{theme}{a ggplot2 theme}
}
\value{
Character string "theme_XXX(...)" or "theme_XXX(...) + theme(...)"
}
\description{
Convert an complete \code{theme} into code to create it
}
