% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-coordinates.R
\name{reverse_coordinates}
\alias{reverse_coordinates}
\title{Convert plot \code{coordinates} into code to create them}
\usage{
reverse_coordinates(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "coord_XXX(...)"
}
\description{
Convert plot \code{coordinates} into code to create them
}
