% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_element}
\alias{reverse_element}
\title{Convert theme \code{element} into code to create it}
\usage{
reverse_element(element)
}
\arguments{
\item{element}{a ggplot2 theme element}
}
\value{
Character string "element_XXX(...)"
}
\description{
Convert theme \code{element} into code to create it
}
