% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet}
\alias{reverse_facet}
\title{Convert plot \code{facet} into code to create it}
\usage{
reverse_facet(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "facet_XXX(...)" or NULL
}
\description{
Convert plot \code{facet} into code to create it
}
