% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet_grid}
\alias{reverse_facet_grid}
\title{Convert plot \code{facet_grid} into code to create it}
\usage{
reverse_facet_grid(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "facet_grid(...)"
}
\description{
Convert plot \code{facet_grid} into code to create it
}
