% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet_labeller}
\alias{reverse_facet_labeller}
\title{Convert facet \code{labeller} argument into code to create it}
\usage{
reverse_facet_labeller(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string: "labeller = label_XXX" or full labeller function body
}
\description{
Try to match the current labeller against all the ggplot2 labellers.  If
no match is found, then just include the full labeller function call
}
