% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet_scales}
\alias{reverse_facet_scales}
\title{Convert facet \code{scale} argument into code to create it}
\usage{
reverse_facet_scales(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string: "scales = '...'"
}
\description{
Convert facet \code{scale} argument into code to create it
}
