% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet_space}
\alias{reverse_facet_space}
\title{Convert facet \code{space} argument into code to create it}
\usage{
reverse_facet_space(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string: "space = '...'"
}
\description{
Convert facet \code{space} argument into code to create it
}
