% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-facet.R
\name{reverse_facet_wrap}
\alias{reverse_facet_wrap}
\title{Convert plot \code{facet_wrap} into code to create it}
\usage{
reverse_facet_wrap(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "facet_wrap(...)"
}
\description{
Convert plot \code{facet_wrap} into code to create it
}
