% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-plot.R
\name{reverse_ggplot}
\alias{reverse_ggplot}
\title{Convert plot \code{ggplot} call into code to create it}
\usage{
reverse_ggplot(p, ggplot_data_name)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{ggplot_data_name}{ggplot data name. Used if identical data match cannot
be found in the \code{plot_env}}
}
\value{
character string "ggplot(...)"
}
\description{
Convert plot \code{ggplot} call into code to create it
}
