% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-plot.R
\name{reverse_ggplot_data}
\alias{reverse_ggplot_data}
\title{Convert plot \code{data} argument into code to create it}
\usage{
reverse_ggplot_data(p, ggplot_data_name)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{ggplot_data_name}{name of primary data source}
}
\value{
character string "data = ..."
}
\description{
Convert plot \code{data} argument into code to create it
}
