% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_incomplete_theme}
\alias{reverse_incomplete_theme}
\title{Convert an incomplete \code{theme} into code to create it}
\usage{
reverse_incomplete_theme(theme)
}
\arguments{
\item{theme}{a ggplot2 theme element}
}
\value{
Character string "theme(...)"
}
\description{
Convert an incomplete \code{theme} into code to create it
}
