% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-labels.R
\name{reverse_labels}
\alias{reverse_labels}
\title{Convert plot \code{labels} into code to create them}
\usage{
reverse_labels(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "labs(...)" or NULL
}
\description{
Convert plot \code{labels} into code to create them
}
