% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer}
\alias{reverse_layer}
\title{Convert plot \code{layer} into code to create it}
\usage{
reverse_layer(layer, p, layer_data_name)
}
\arguments{
\item{layer}{ggplot2 layer object}

\item{p}{ggplot2 plot object}

\item{layer_data_name}{name}
}
\value{
character string "geom_XXX(...)"
}
\description{
Convert plot \code{layer} into code to create it
}
