% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer_aes_params}
\alias{reverse_layer_aes_params}
\title{Convert layer \code{aes_params} into code to create them}
\usage{
reverse_layer_aes_params(layer)
}
\arguments{
\item{layer}{ggplot2 layer object}
}
\value{
character string "size = 3, na.rm = TRUE, ..." or NULL
}
\description{
Convert layer \code{aes_params} into code to create them
}
