% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer_data}
\alias{reverse_layer_data}
\title{Convert layer \code{data} into code to create it}
\usage{
reverse_layer_data(layer, p, layer_data_name)
}
\arguments{
\item{layer}{ggplot2 layer object}

\item{p}{ggplot2 object}

\item{layer_data_name}{default name to use for data if it isn't waiver, and
matching data cannot be found in the \code{plot_env}}
}
\value{
character string of format "data = ..." or NULL
}
\description{
Convert layer \code{data} into code to create it
}
