% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer_mapping}
\alias{reverse_layer_mapping}
\title{Convert layer \code{mapping} argument into code to create it}
\usage{
reverse_layer_mapping(layer)
}
\arguments{
\item{layer}{ggplot2 layer object}
}
\value{
character string "mapping = aes(x = ..., y = ...)" or NULL
}
\description{
Convert layer \code{mapping} argument into code to create it
}
