% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer_position}
\alias{reverse_layer_position}
\title{Convert layer \code{position} into code to create it}
\usage{
reverse_layer_position(layer)
}
\arguments{
\item{layer}{ggplot2 layer object}
}
\value{
character string "position = position_XXX(x, y)"
}
\description{
Convert layer \code{position} into code to create it
}
