% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layer_stat}
\alias{reverse_layer_stat}
\title{Convert layer \code{stat} into code to create it}
\usage{
reverse_layer_stat(layer)
}
\arguments{
\item{layer}{ggplot2 layer object}
}
\value{
character string of format "stat = 'statname'"
}
\description{
Convert layer \code{stat} into code to create it
}
