% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-layer.R
\name{reverse_layers}
\alias{reverse_layers}
\title{Convert all plot \code{layers} into code to create them}
\usage{
reverse_layers(p, layer_data_names)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{layer_data_names}{names}
}
\value{
vector of character strings - one for each layer
}
\description{
Convert all plot \code{layers} into code to create them
}
