% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_margin}
\alias{reverse_margin}
\title{Convert a theme \code{margin} into code to create it}
\usage{
reverse_margin(theme_margin)
}
\arguments{
\item{theme_margin}{theme margin item}
}
\value{
character string "margin(...)"
}
\description{
Convert a theme \code{margin} into code to create it
}
