% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-scales.R
\name{reverse_scales}
\alias{reverse_scales}
\title{Convert plot \code{scales} into code to create them}
\usage{
reverse_scales(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
character string "scales_XXX(...)"
}
\description{
Convert plot \code{scales} into code to create them
}
