% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_theme}
\alias{reverse_theme}
\title{Convert plot \code{theme} into code to create it}
\usage{
reverse_theme(p)
}
\arguments{
\item{p}{ggplot2 plot object}
}
\value{
Character string "theme_XXX(...)" or "theme(...)" or NULL
}
\description{
For incomplete themes, produce an explicit \code{theme()} call with all
theme arguments.
}
\details{
Otherwise, try and match against the standard ggplot themes.
}
