% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_theme_item}
\alias{reverse_theme_item}
\title{Convert a theme \code{item} into code to create it}
\usage{
reverse_theme_item(theme_item)
}
\arguments{
\item{theme_item}{a ggplot2 theme item}
}
\value{
Character string "element_text(...)", "margin(...)" etc
}
\description{
Theme items can be:
\itemize{
\item{element objects}
\item{margin objects}
\item{unit objects}
\item{character strings}
\item{logical values}
\item{numeric values}
}
}
