% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{reverse_unit}
\alias{reverse_unit}
\title{Convert a theme \code{unit} into code to create it}
\usage{
reverse_unit(theme_unit)
}
\arguments{
\item{theme_unit}{theme unit item}
}
\value{
character string "unit(...)"
}
\description{
Convert a theme \code{unit} into code to create it
}
