% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-theme.R
\name{shallow_diff}
\alias{shallow_diff}
\title{Shallow diff between 2 named lists}
\usage{
shallow_diff(ref, test)
}
\arguments{
\item{ref}{reference list of defaults}

\item{test}{test list including customisations}
}
\value{
named list of just the customisations which differ from the reference
}
\description{
Shallow diff between 2 named lists
}
