% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snakeize}
\alias{snakeize}
\title{Local copy of 'snakeize' function that isn't exported from 'ggplot2'}
\usage{
snakeize(x)
}
\arguments{
\item{x}{geom or stat name in CamelCase e.g. "GeomPoint"}
}
\value{
name in "snake_case" e.g. "geom_point"
}
\description{
Local copy of 'snakeize' function that isn't exported from 'ggplot2'
}
